requirements.txt
# AI API Generator - Complete Dependencies

# Main Application Dependencies
streamlit==1.31.0
requests>=2.31.0
python-dotenv>=1.0.0  # For loading .env files

# Universal Scraping Layer Dependencies
# Core HTTP and HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
cssselect>=1.2.0
html5lib>=1.1

# Async HTTP client
aiohttp>=3.9.0

# Browser automation (for dynamic content)
playwright>=1.40.0

# System monitoring
psutil>=5.9.0

# API Server Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0

# Caching backends (optional)
redis>=5.0.0

# Configuration file support
pyyaml>=6.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.88.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development Dependencies (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0